Add virt-xml sidecar hook
The virt-xml tools is a command line utility for editing libvirt XML.

This setup introduces a new hook to integrate virt-xml in KubeVirt.
Using virt-xml allows to use a well-maintained tool to try and test unsupported
options by the KubeVirt APIs.

This feature is supposed to help developers to try and test unsupported
options without the need to develop their own sidecar.

Signed-off-by: Alice Frosi <afrosi@redhat.com>